Updated Validate Function
Reworked the validate function to validate using the jQuery validate
plugin. Added missing name attributes to the HTML tags. Added the
missing class attributes to the HTML tags that are required form
elements. Changed the color of the "error" text generated to red and
5px left margin. Corrected an error with the date input that prevented
it from being selected; the type is now "text". Corrected a typo in the
script file declaration of the validator, validate now works (since
it's now included).