Finalized jQuery switch
I corrected the checkbox issue. Replaced all .checked methods to the
jQuery .prop("checked", true) for setting and .prop("checked", false)
for unsetting. Fixed a recurring issue with my date field not disabling
and renabling by using the jQuery .removeAttr() method instead of
setting the attribute to an empty string. Fixed an error with the
radiobox mass setting loop. The array needed to be wrapped in the
jQuery factory. New line reads like this: $(tags[i]).prop("checked",
false). Added a call to the reset data function after the edit routine
function is called to clear any lingering form fields. Added an extra
validation for the date because I want only new dates entered not old
dates. Encountered an error with the localStorage key not passing
correctly or not at all due to the use of this. Cleared that this.key
and passed only a key value that is retained through the use of an
attribute called key saved in each of the editDelete routine Li's. When
the links are made the proper key is retrieved and passed into the
binded event. Replaced the .value with the .val() method. Fixed an
issue with the image files not displaying properly. Cleared the old way
of retrieving the file name for a shorter simpler method. The file name
is now derived from index 0 to first white space, which is similar to
how it previously worked. Removed old code that was commented out.