Date error
Corrected an error that occurs with the date in iPhone. The date format
won't recognize the old formatting that used dashes as separators but
accepts slashes instead. So forward slash marks have replaced the
dashes. Trying to resolve the ugly look that jQuery uses when
validating. Namely injecting labels with the required messages. The
problem is it appends after the input and breaks the look of the
layout. I'm experimenting with absolute positioning to resolve this.
Replaced the window.location.reload() with location.reload(). Replaced
the $(document).ready() with the recommended document.bind("pageinit").
Update the date values in the JSON data to reflect the new slash
instead of dash date formats. Update the index.html date field
placeholder and input text also for the same reason.