jQuerified checkboxes and radio buttons
Refactored my 'getCheckboxValue' and 'setCheckboxValue' functions to
completely use jQuery methods. Replaced my lengthy loops based on
hardcoded property names with jQuery logic. Property names are now
dynamically retrieved and used to get and set the values. Created a
function that contains loops to refresh the select, checkbox, radio and
slider values. Corrected an error that continuously set my date field
to disabled using .attr(). Further refactored that to use the jQm
method of .textinput() to disable and enable the date field. Removed
the resetData function and added it's different funtionality to the
restoreDefault function. Added conditionals to determine what is being
reset or defaulted. Fixed an error that occurred when restoreDefault
function was called because of the element refreshing was occurring
before the page loaded. Added a bound event to the add routine page
that will restore the defaults after it is loaded to correct the issue.
Used jQuery logic to refactor my 'setInlayLinks' function which used
hardcoded property values to loop through and assign click events. Now
the ids are retrieved using jQuery method calls instead, then bound to
a click event. Updated my search bar to use jQuery search type. Removed
the functionality for now. Removed more commented out and useless
functions.