Update index file & css
Redesign some of the html elements in the index file. Wrapped the
checkboxes and radio buttons in div controlgroups to give jQuery look.
Fixed the error with the checkbox and radio buttons now receiving that
jQ styling because they were wrapped in <p> elements. Styled the
listinlay on the home page with data-role of "button". Had the same
issue with the syling not taking place because they were wrapped in <p>
also because the types were capitalized. Updated my file information:
date, course, etc