Added data interchange functionality
Added a dynamic jQuery ajax function for sending httprequests. Function
is customized to build the ajax object from passed in arguments or will
accept a predefined ajax object. Added 3 custom functions for parsing
my external XML, CSV, and JSON data. The parseXML function loads the
xml file from the ajax request and appends the child nodes to a div in
the document. Loops through the elements and creates a new parsed
object of that data. The parseCSV function loads the text data from the
ajax request and using logic, splits the data by new line characters,
commas, and spaces. The parseJSON function takes the json information
from the external file and parses it using the built in methods. Added
bound events to the new data interchange buttons in the show data part
of the form. When clicked the events create an ajax request object
tailored to the data the button represents (json, csv, xml). Changed
the bound event of the show data button to toggle the routine list /
form view instead of the getData function call since the new data
buttons will do this. This is required to give the user a choice in the
type of data to load instead of just loading JSON. Removed the use of
the multiplication symbol as signifiers as this created problems in the
CSV. The signifier for the days of the week is now just an "X".