Resolved storage issue
Fixed the localStorage issue in safari. Safari browser adds system data
to the localStorage. This caused issues with my checking for data
inside of storage. Also created problems because the system data would
be loaded and break my code. Created a function called killInvalidLS
(kill invalid local storage). The function checks for non-json object
data strings in local storage. If found that particular item will be
removed leaving only my data. Corrected an issue with the killInvalidLS
function not operating in Firefox by looping through the storage
getting by item. Added confirmation prompts on the loading of the data
interchange formats. Added functionality that removes any data in
storage and loads only the format selected. Thus the reason for the
confirmation. Corrected date issue in iPhone. (hopefully). Made file
wide change to the camel case key value names associated with item
data. This is to correspond with the changes to the XML, CSV, and JSON
data mentioned earlier. Corrected the toggle controls issue (missing an
argument)