Updated stylesheet
Stylesheet contained old and obsolete references and styles from MIU.
Removed all commented out sections along with old styles that I no
longer employ such as the custom drop down in the getData function
call. Now replaced with the jQuery accordion. Updated styles that
referred to elements of the same id to refer to elements with that
class instead for HTML5 validation. Added styling to the getData images.