Fixed validation errors
Corrected validation errors: fixed duplicate element id's used for
styling to instead be classes, fixed markup typos that included missing
h in a href attribute (was ref), fixed a type that called the name
attribute 'hname'. Probably happened in an copy / paste function.
Removed the HTML5 'button' element and instead styled the anchor as a
data-role='button'. Added a hash symbol in the form post was an empty
string. Changed the data interchange format buttons to be inputs of
type 'button' inside of a 3 column jQuery mobile grid. Removed some
hardcoded style attributes to instead be styled using the stylesheet.