jQuerify getData list & refactoring
Removed more obsolete functions and or code. Added alt property to the
dynamic img creation for HTML5 validation. Reworked the getData list to
be an accordion of listviews of data and removed the old custom format.
Added the routine images to the accordion header instead of their own
list, for use as icons. Fixed an error where the accordion would not
add the jQuery styling by triggering the "create" event on the
collapsible-set. Reworked the 'edit' and 'delete' links to be styled as
jQuery buttons in an 2 column jQuery grid. Rearranged the binding
scripts and added more comments to them for clarity. Changed the
document.bind("pageinit") back to .ready(). Trying to resolve iOS XML
error.