Update App look
Removed the nav element on the "myRoutine" page and moved the
functionality to buttons in the footer. Created a div in the footer of
data-role="navbar" to house a list of the old nav anchors. jQuery
styles these as buttons automatically. Removed sytling for those old
nav buttons. Removed remaining old css styling for list items in the
getData function.