Fixed iOS XML problem
Updated the toggle function to reflect the changes made to the
navigation. Removed the button toggling methods .hide() and .show()
since the buttons are now in the footer. Instead they anchors will be
toggled to be disabled or not to prevent the functionality when it
shouldn't be. This leaves the buttons visible and fixes an error that
happens when the nav buttons are hidden then re-shown. To distinguish
to the user where they are on the routines page the toggle will also
change the link color associated with the current configuration
otherwise it will be left white. Removed old alerts and commented out
sections in the getData function from the code. Added a killInvalidLS
function check in the "confirmAutoFill function. And finally I fixed
the error that prevented the XML from being loaded and parsed in iOS!
(yeaaaaaaah!!!)  I was appending the parsed XML child nodes to an empty
page and iOS wasn't having it. Removed the old appending method and
instead use a jQuery .each() method to return an array of the parent
elements. From there my previous code picks up and cycles through that
array parsing out the data and assigning it to the proper object key
value. Stores it in the xml variable and voila!