Commit stylesheet and couchApp files
Created a slimmed down stylesheet using the stylesheet from the
gh-pages for consistency. Changed the couchDB app _id file to
"_design/fitapp". Changed the couchapp.json file to reflect the app
name and description. Set up the local and cloudant branches for my
couchDB app. Created the cloudant account and gave permissions to the
mdv account.