Commit my couchDb views
Created six views for my couch app. One for each kind of exercise
routine in the json and a sixth view for viewing all of the routines.
Created a javascript global logic to use for the conditional instead of
the counting characters method. A variable is created which stores the
index value of the colon in the _id. The logic then uses that colon
variable to compare for the right routine type and also for emitting
the shortened id variant.