Committed CouchDb JS file
Created a javascript file for the couchDb index file. Copied my newXHR
function over from my gh-pages app script file for the functionality.
Created a function called "makeLists". This function is similar to my
"autoFill" function in the gh-pages app, in that it is responsible for
creating the proper ajax request. By default it will always show all
data. Created a function called "parseJSON". This function is a
streamlined version of the "getData" function designed especially for
task of looping through the couchDB app object and creating an
accordion of listviews based on the type of filter selected. Lastly, I
bound an anonymous function that calls the "makeLists" function to each
of the accordion buttons on the home page. The logic passes in the
anchor's id which is tested in a switch case method inside the
"makeLists" function. Linked the cta image anchor to an anonymous
function that calls the "makeLists" function but passes in the
"showall" argument and not the "this.id".