Added my scripts
Trimmed the news stream function from the scrip file. Refactored the
getCheckbox and setCheckbox functions to use a for loop to further
optimize the setting. Added a getExerciseClass function. This function
will determine the type of exercise based on the routine set. I added
this new object key to help me sort the couchDB result object because I
am now using the couch id keys. Added the ability to save, edit, and
delete data from the database. I also chose to keep the localStorage
for this purpose. When the data is fetched it is parsed and then stored
into localStorage. The getData function can then load the localStorage
data without having to be refactored. Since I'm using two different
types of storage in my app I need to save both keys for use. When the
data is fetched the _id and _rev keys are stored into an array and
passed along with the localStorage key. When to load the proper data
from storage the localStorage key is passed in to populated the form
fields. But the couchDB _id and _rev keys are passed into the form
validation and saving function calls, since the localStorage data is
never saved just loaded dynamically from the couchDB and displayed.
Added a new function called getDB() which accepts some arguments and
pulls the data based on those arguments. Added the "pageshow" handler
for the new routines.html page. When it shows the argument passed in is
captured and used to display the proper view.