Added my db views
Refactored the views as I am no longer using custom IDs. I had to add a
"type" key to the routine objects, and I use that key to sort through
the views. Because I need to use the _id and _rev keys I also added
them to the list of data to be emitted.