Fixed database key error and page reload problem
Removed the old refreshing method after an item is deleted. Replaced it
with functionality that clears the old list and fetches the new data
for display without changing the page. Fixed an error that would occur
when trying to fetch the url for the list refresh because the anchor
was linking to the routine page. The link now links to the list view
page with the view that was previously selected. That view is now
stored in the anchor href and the url extractor can properly fetch that
data. Fixed an error that happend after the first save or edit had
occurred. All subsequent edits or saves would not reflect the update
IDs and Revision values. Re-factored the way that database key was
passed around. The key is no longer passed into the validation methods
but instead set into a key attribute on the save button in the edit
routine function. When the date is validated the key is fetched, split
by the comma, and passed into the storeData function call. No editing
was required of the storeData. Added the code to remove the key from
the save button when the defaults are reset. Otherwise the data would
try to edit a routine instead of saving a new one. Deleted the old
routines.js script.